+++ CHANGES.md
@@ -15,6 +15,7 @@
 - Fixed a fault that prevented the cached window object being used to open new windows.
 - Added "Visible Days" tab to the Configuration dialog, allowing the visible days in the display to be toggled.
 - Added new option "urlWindowTarget", which states the target for a event Url that is opened (defaults to _blank for a new window).
+- Added "Enable day name headers in the main display" option to "Display" on the Configuration dialog.
 - 
 
 ### Version 0.9.9: